Merged PR 15793: Use one image per partition and update fstab in reconcile
This PR is a followup to [#15611](https://dev.azure.com/mariner-org/ECF/_git/trident/pullrequest/15611). It changes the handling of images so that each one applies to a single partition and demonstrates `Module::reconcile` by using the hook to implement updates to /etc/fstab.

Right now this is still using partition labels for identifying partitions. However, that's somewhat brittle so it would be good to switch to a different option.

The host config now looks something like:
```yaml
storage:
  disks:
    - device: "/dev/disk/by-path/pci-0000:00:1f.2-ata-1.0"
      partition-table-type: gpt
      partitions:
        - type: esp
          size: 1G
        - type: root-a
          size: 20G
        - type: root-b
          size: 10G
imaging:
  images:
    esp:
      url: "http://10.137.192.130:8088/boot_partition.raw.zst"
      sha256: b789802a93269486dc070faad41dc4d3bf0c5d018d3dea772917cf457a7f3115
      format: raw-zstd
    root:
      url: "http://10.137.192.130:8088/root_fs_partition.raw.zst"
      sha256: 8c11d8fe3a99d360f204e789d426ed70aaf4bf72163c8f4476a31fcf894601e6
      format: raw-zstd
network:
  ethernets:
    vmeths:
      match:
        name: enp*
      dhcp4: true
```

Related work items: #5560