Merged PR 15796: Update dependency versions
A bunch of dependencies have released new patch or minor versions. This PR upgrades our Cargo.toml and Cargo.lock to use them.

Specifically I ran `cargo update` and `cargo upgrade`, the latter of which requires [cargo-edit](https://github.com/killercup/cargo-edit) to be installed.

It is probably possible to automate updates like this, but not sure whether it is worth the effort

Related work items: #5560