Merged PR 15932: Persist host status
This PR adds code to use sqlite to save the host status after each operation during provisioning / update, and does a bunch of refactoring around mounting and unmounting partitions.

In a followup, I'll make the changes necessary for trident to start on the runtime OS and confirm that loading the saved host status works.

Related work items: #5561