Merged PR 16764: Fixup var for sshd
Without the fix, sshd will fail to start if you mount a custom volume over /var. That is because sshd expects /var/lib/sshd to be already present. This PR fixes that by manually creating it if it is missing. Longer term, we should join forces with MIC to migrate files from the root volume.

Related work items: #6032