Merged PR 16871: Fix: SSHD Config
Fix: SSHD Config

The users submodule was not checking if sshd was present before trying to overwrite settings. We now check to make sure it exists before trying to set sshd settings.

The user submodule was not making sure that the arrays are not empty for each targetted setting, hence in some scenarios it would produce invalid configs with no specific targets. We now check that we actually need to write those rules before we try to write them.

A HC with with default ssh settings will result in the following:
![image.png](https://dev.azure.com/mariner-org/2311650c-e79e-4301-b4d2-96543fdd84ff/_apis/git/repositories/895b6b3d-5077-488a-8001-ab6b5a14c1a3/pullRequests/16871/attachments/image.png)
(When SSHD is present)

Related work items: #5993