Merged PR 16887: Adding workDirectory so steps template can be used in other repo pipelines
Adding parameter for the working directory that tasks will use. The default value is `$(Build.SourcesDirectory)`, which would be the trident repository in a pipeline that only uses the code of one repository (trident).

Paths used had been updated to use the new parameter and workingDirectory value had been set for tasks that use the content of trident repository.

Example of use the of trident template in argus-toolkit Onebranch pipeline:
https://dev.azure.com/mariner-org/ECF/_build/results?buildId=447352&view=results

Related work items: #6025