Merged PR 16549: Basic e2e Flow of A/B Upgrade with Trident
This PR implements a basic A/B update flow via systemd-sysupdate with Trident. The users are able to update the root partition; other types of partitions will be eligible for A/B update in a later iteration. The logic is implemented mostly inside the Systemd-Sysupdate sub-module of the Image module, in systemd-sysupdate.rs. This sub-module writes a transfer config file based on HostConfiguration provided by the user and communicates with the systemd-sysupdate executable using that file. The file named previously image.rs has been renamed to image/mod.rs. The code responsible for direct streaming of images in image/mod.rs has been refactored into a separate sub-module, inside of image/stream_image.rs. Utility functions have been also added to lib.rs.

The optional ab-update section did not change from the previous iteration, and it contains the same fields. The details on how to complete a successful A/B update flow with Trident is described in the Get Started with Systemd-Sysupdate section in README.md, along with sample HostConfiguration files.

Related work items: #5273, #6150