Merged PR 16729: Fix get_update_kind for image.rs for abvolumepair
Without this patch, post A/B update reboot, get_update_kind finds image discrepancy, because the invalid volume of the abupdatevolume pair is checked. This patch updates the check to not assume any update is in progress and check existing active volumes instead. And for that to work, the reconcile state Ready will return the active volume.

In addition, adds some extra UTs and moves helper function closer to they usage.

Related work items: #5273