Merged PR 17422: Persist datastore to temporary location rather than keeping in memory
Instead of using an in-memory datastore that is placed onto disk once provisioning has completed, this PR changes things to always place the datastore on the filesystem and merely relocate it once the provisioning step is done.

It also delays closing the datastore and produces clearer errors if further attempts are made to use the datatsore once it is closed.

Related work items: #6145