Merged PR 17472: Fix: Use correct private key in functional tests
We were passing the public key to paramiko where it expects the private key. It sill worked because paramiko will look for keys by default, so it found the right one in `~/.ssh/` (if it was there).
This change makes it so the intended private key is passed.

Related work items: #6474