Merged PR 17507: Allow overriding safety-check using a marker file
Trident tries to avoid accidentally overwriting a dev machine or other Linux box by performing some safety checks on its environment. Currently that involves checking whether the system is booted from a ramdisk, which is very robust but interferes with certain customer scenarios.

This PR adds an additional way to override the check: placing a file at the root of the filesystem indicating that the user is OK with potentially destroying everything on the system. Remember, the first step of trident provisioning basically amounts to running `rm -rf /`!

Related work items: #6145