Merged PR 17649: Fix trident error Unsupported AB volume selection when AB not configured
Trident errors with 'Unsupported AB volume selection when AB not configured' in set_boot_entries() when AB is not configured in host configuration. This is due to get_ab_update_volume() returning None early with `let active_volume = &host_status.storage.ab_update.as_ref()?.active_volume`. The fix is to refactor the code such that storage.ab_update is only referenced when we are sure it is configured.

Related work items: #6529