Merged PR 17687: Record location of persistent datastore on provisioning OS
This PR changes trident to record the block device and relative path of the datastore in a `/var/lib/trident/datastore-location.conf` file on the provisioning OS's filesystem. It eliminates the need to copy the datastore back to the provisioning OS (which was previously only happening if provisioning _succeeded_).

Related work items: #6145, #6471