Merged PR 17718: docs: Fixing image names in sample HC for esp and root files
<!-- Please, remember to prefix the title with one of:
 - `bug:`
 - `feature:`
 - `engineering:`
 - `docs:`.
For details, check [Pull Requests](/dev-docs/pull-requests.md) and
[this page](https://dev.azure.com/mariner-org/ECF/_git/ecf-docs?version=GBmain&path=/TeamDocs/dev-guide.md&_a=preview). -->

# 🔍 Description
The sample HC that we produce shows the runtime partitions as `boot.raw.zst` and `root.raw.zst` but trident will fail to find the image if we have a HC that is following the sample because the `make download-runtime-partition-images` in argus-toolkit renames the files to `esp.rawzst` and `root.rawzst`. This change makes the sample HC functional.

# 🤔 Rationale
<!-- (can be skipped for bug fixes)
Why is this PR needed?
Reviewers should get a clear understanding of the goals of this PR.
-->
Have a sample HC with values that work with trident's expectations.

# 📝 Checks
<!-- Just a quick reminder :) -->

- [X] Format Check
- [X] Check rust
- [X] Updated API docs
- [X] Tests Pass
- [X] Functional Tests Pass
- [X] Local make targets work
  - [X] `make rpm`
  - [X] `make docker-build`

Related work items: #6588