Merged PR 17714: feature: Additional running in a container check
# 🔍 Description
Adding additional check to see if Trident is running in a container, in case the container has not be correctly configured with the explicit configuration that we expect. This is to help users detect they are running with unsupported configuration and fix that.

# 🤔 Rationale
This will help users detect why the system is misbehaving.

# 📝 Checks
- [x] Format Check
- [x] Check rust
- [x] Tests Pass
- [x] Functional Tests Pass
- [x] Local make targets work
  - [x] `make rpm`
  - [x] `make docker-build`

Related work items: #6267