Merged PR 17405: feature: Implement file-based update of ESP
# 🔍 Description
<!--
What is this PR about?
Reviewers should get a clear understanding of what is being done/changed.
-->
This PR implements the following:

1. While deploying new images requested by HostConfiguration, Trident downloads the ESP partition image directly to the disk, instead of writing it to a block device.
2. Trident mounts the downloaded image to a temp dir and searches for the three keys files, copying each into the corresponding ESP directory, either /EFI/azlinuxA or /EFI/azlinuxB.

# 🤔 Rationale
<!-- (can be skipped for bug fixes)
Why is this PR needed?
Reviewers should get a clear understanding of the goals of this PR.
-->
To implement firmware reboot in Trident, it is necessary to first implement file-based updates of the ESP partition. Per current layout, there will be two disks, with a single ESP partition on each. Each ESP partition will have two copies, A and B, of the key files required for booting: bootx64.efi, grubx64.efi, and grub.cfg. Each set of files will be placed into a separate directory, /EFI/azlinuxA or /EFI/azlinuxB. Following the same logic as with block-based A/B upgrades, while performing a file-based upgrade of ESP as part of StageUpdate, Trident will copy the files from the new ESP image into the corresponding directory:
1. If volume A is currently active, copy boot files into /EFI/azlinuxB,
2. If volume B is currently active OR no volume is currently active, i.e., Trident is doing CleanInstall, copy boot files into /EFI/azlinuxA.

As part of Update, Trident will set the UEFI firmware variables, to ensure that the firmware boots into the updated partition, while using the “old,” currently active, partition as the fallback option. If the system is currently running from A, Trident will use the temporary boot device logic to direct the firmware to first search for the bootable device B, where the image has been updated as part of StageUpdate. If the reboot succeeds, Trident will then update the boot order to ensure that on the next reboot, from B back to A, the firmware will boot into B as the fallback mechanism. Otherwise, if the reboot into device B fails, the firmware will try to boot from the next bootable device, which is A. As a result, on a failed reboot, the firmware will perform a rollback, as outlined in the next bullet point.

Related work items: #6383