Merged PR 17762: engineering: Changing download from latest to latest from branch
<!-- Please, remember to prefix the title with one of:
 - `bug:`
 - `feature:`
 - `engineering:`
 - `docs:`.
For details, check [Pull Requests](/dev-docs/pull-requests.md) and
[this page](https://dev.azure.com/mariner-org/ECF/_git/ecf-docs?version=GBmain&path=/TeamDocs/dev-guide.md&_a=preview). -->

# 🔍 Description
<!--
What is this PR about?
Reviewers should get a clear understanding of what is being done/changed.
-->
Changing in downloading artifacts the type of download from latest to latestFromBranch, so it uses main branch, as apparently latest just ignore the branch specified.

# 🤔 Rationale
<!-- (can be skipped for bug fixes)
Why is this PR needed?
Reviewers should get a clear understanding of the goals of this PR.
-->
We want to download the latest from the main branch, not the latest from any branch that succeeded.

Related work items: #6560