Merged PR 17775: bug: Regenerate initrd to unblock boot
# 🔍 Description
Regenerate initrd for every run, to both unblock the new A/B update code and RAID e2e. This is needed due to the change in the images to no longer regenerate initrd as part of image build, which was breaking RAID deployments. This is a stop gap solution to unblock pipelines, until we better understand what broke the RAID e2e.

# 🤔 Rationale
Unblock pipelines.

Related work items: #6638