Merged PR 17794: bug: Only run official nightly
# 🔍 Description
trident-official is currently running with every change on main, but we only want to run it nightly. This change should do that.

# 🤔 Rationale
We are already running trident on merges to main, so this is not needed. Also, official is doing publishing, and we would rather decouple merge from publishing, so we can and fix any additional issues.

Related work items: #6670