Merged PR 17779: Recursively unmount /mnt/newroot at startup so trident can rerun on failure
<!-- Please, remember to prefix the title with one of:
 - `bug:`
 - `feature:`
 - `engineering:`
 - `docs:`.
For details, check [Pull Requests](/dev-docs/pull-requests.md) and
[this page](https://dev.azure.com/mariner-org/ECF/_git/ecf-docs?version=GBmain&path=/TeamDocs/dev-guide.md&_a=preview). -->

# 🔍 Description

If the /mnt/newroot directory already exists when Trident starts provisioning storage, recursively unmount it. Also adds a functional test for recursive unmount and an E2E tests with a post-provision script that crashes the clean-install process the first time it is run, but not subsequent times.

# 🤔 Rationale

Ensure that trident leaves the system in a state where it can run again without having to do a reboot.

# 📝 Checks
<!-- Just a quick reminder :) -->

- [x] Format Check
- [x] Check rust
- [x] Updated API docs
- [x] Tests Pass
- [x] Functional Tests Pass
- [x] Local make targets work
  - [x] `make rpm`
  - [x] `make docker-build`

Related work items: #6145, #6435, #6566