Merged PR 17828: docs: Add how-to guide for offline validator
<!-- Please, remember to prefix the title with one of:
 - `bug:`
 - `feature:`
 - `engineering:`
 - `docs:`.
For details, check [Pull Requests](/dev-docs/pull-requests.md) and
[this page](https://dev.azure.com/mariner-org/ECF/_git/ecf-docs?version=GBmain&path=/TeamDocs/dev-guide.md&_a=preview). -->

# 🔍 Description
<!--
What is this PR about?
Reviewers should get a clear understanding of what is being done/changed.
-->

Add how to guide about the offline HC validator.
Also clean up an error message that was printing badly formatted redundant info.

# 🤔 Rationale
<!-- (can be skipped for bug fixes)
Why is this PR needed?
Reviewers should get a clear understanding of the goals of this PR.
-->
Document how to use this feature.

# 📝 Checks
<!-- Just a quick reminder :) -->

- [X] Format Check
- [X] Check rust
- [X] Updated API docs
- [X] Tests Pass
- [X] Functional Tests Pass
- [X] Local make targets work
  - [X] `make rpm`
  - [X] `make docker-build`

# 📌 Follow-ups
<!-- When applicable, add links to tasks or PR for work that will be addressed
later as a follow up of this PR. -->

# 🗒️ Notes
<!-- Add any other relevant info -->

Related work items: #6614