Merged PR 17882: engineering: Add tasks to build PXE files
This is a somewhat hacky approach to publish the files needed for PXE booting.

The general strategy is to take the generated installer ISO and extract the relevant pieces from it, with a few items downloaded from PMC instead. Then the relevant files are packaged back up into a ZIP file that is expected to be extracted and served from a TFTP server.

Related work items: #6743