Merged PR 17912: engineering: Name to distinguish e2e_tests results
# 🔍 Description
Names for e2e_tests results include now Stage name and again the configuration and job attempt number.

# 🤔 Rationale
Notice that the name of e2e_tests results had removed the differentiators. This is probably, because of a bad merge when the template for the e2e_tests was introduced. Adding stage name as now we run the same tests for BM.

Related work items: #5966