Merged PR 17916: engineering: cleanup prov-os build
<!-- Please, remember to prefix the title with one of:
 - `bug:`
 - `feature:`
 - `engineering:`
 - `docs:`.
For details, check [Pull Requests](/dev-docs/pull-requests.md) and
[this page](https://dev.azure.com/mariner-org/ECF/_git/ecf-docs?version=GBmain&path=/TeamDocs/dev-guide.md&_a=preview). -->

# 🔍 Description
<!--
What is this PR about?
Reviewers should get a clear understanding of what is being done/changed.
-->

Cleanup prov-os build to align with changes in Argus.
Required by !17915

# 🤔 Rationale
<!-- (can be skipped for bug fixes)
Why is this PR needed?
Reviewers should get a clear understanding of the goals of this PR.
-->

The pipeline needlessly depended on argus's makefile.

# 📝 Checks
<!-- Just a quick reminder :) -->

- [ ] Format Check
- [ ] Check rust
- [ ] Updated API docs
- [ ] Tests Pass
- [ ] Functional Tests Pass
- [ ] Local make targets work
  - [ ] `make rpm`
  - [ ] `make docker-build`

# 📌 Follow-ups
<!-- When applicable, add links to tasks or PR for work that will be addressed
later as a follow up of this PR. -->

# 🗒️ Notes
<!-- Add any other relevant info -->

input to artifacts

Related work items: #6764