Merged PR 17840: engineering: Test Makefile during PR checks
# 🔍 Description

Validate the Makefile before a PR is merged, rather than only after.

# 🤔 Rationale

We currently have a bunch of steps listed in the PR template that developers are expected to run manually before merging their PR and then have a pipeline that runs after merging to detect regressions. This is completely backwards! We should instead be using our pipelines to detect possible regressions before PRs are merged

Related work items: #6713