Merged PR 17923: engineering: Helper module for working with grub.cfg
# 🔍 Description
Helper module for reading and writing grub.cfg. This is probably very temporary, as we might be switching to grub-mkconfig and might be able to make more assumptions. Nonetheless, it will help catch some of the issues we have already hit and it centralizes grub config changes, so they can be easily unit tested as well.

Deadcode present in this PR will be used once the rest of the verity enablement is merged.

# 🤔 Rationale
Common tested logic for accessing grub.cfg.

Related work items: #6758