Merged PR 17918: bug: Disable kexec temporarily
<!-- Please, remember to prefix the title with one of:
 - `bug:`
 - `feature:`
 - `engineering:`
 - `docs:`.
For details, check [Pull Requests](/dev-docs/pull-requests.md) and
[this page](https://dev.azure.com/mariner-org/ECF/_git/ecf-docs?version=GBmain&path=/TeamDocs/dev-guide.md&_a=preview). -->

# 🔍 Description
<!--
What is this PR about?
Reviewers should get a clear understanding of what is being done/changed.
-->

Disables use of kexec in favor of hard reboot.

# 🤔 Rationale
<!-- (can be skipped for bug fixes)
Why is this PR needed?
Reviewers should get a clear understanding of the goals of this PR.
-->

Stabilize hard reboot, then stabilize & re-enable kexec.

# 📝 Checks
<!-- Just a quick reminder :) -->

- [X] Format Check
- [X] Check rust
- [X] Updated API docs
- [X] Tests Pass
- [ ] Functional Tests Pass
- [X] Local make targets work
  - [X] `make rpm`
  - [X] `make docker-build`

# 📌 Follow-ups
<!-- When applicable, add links to tasks or PR for work that will be addressed
later as a follow up of this PR. -->

- #6775

# 🗒️ Notes
<!-- Add any other relevant info -->

Related work items: #6774