Merged PR 17941: engineering: Update coverage instructions and bump UT baseline
# 🔍 Description
I noticed that Coverage Gutters is not picking the generated coverage artifacts by defaults. Added notes on how to fix and also bumping the UT baseline based on values captured. As a side effect, removing one of the coverage artifacts (lcov) that nobody is consuming.

# 🤔 Rationale
Coverage instructions should be helpful to get started.

Related work items: #6758