Merged PR 17940: bug: Fixes grub config linux command line update logic
# 🔍 Description
The recently added logic to update grub config kernel command line was using regular expression replace(), which was mishandling the `$` signs in the string to replace and treating them as variables (and resulting in them being removed from the replaced string). The fix is instead using replace_range() logic, and only using regex for consistent matching.

# 🤔 Rationale
We need to expect that linux command line will have some placeholders starting with `$`, and this change fixes it.

Related work items: #6805