Merged PR 17763: feature: Finalize Boot order to support RollBack
<!-- Please, remember to prefix the title with one of:
 - `bug:`
 - `feature:`
 - `engineering:`
 - `docs:`.
For details, check [Pull Requests](/dev-docs/pull-requests.md) and
[this page](https://dev.azure.com/mariner-org/ECF/_git/ecf-docs?version=GBmain&path=/TeamDocs/dev-guide.md&_a=preview). -->

# 🔍 Description
<!--
What is this PR about?
Reviewers should get a clear understanding of what is being done/changed.
-->
1. Trident currently utilizes efibootmgr to configure the BootNext Variable following an AB update. The process involves adjusting the BootOrder based on the success or failure of the update. The BootNext variable, set during the update, is saved in the database.
If booting into the partition corresponding to BootNext is successful, indicating a successful update, the BootOrder is modified to include this entry. Conversely, if the BootNext variable from the database doesn't match BootCurrent, signifying a failed update, the system rolls back to the BootCurrent entry. This approach enables support for rollbacks using efibootmgr boot entries.

2. extended unit tests for updated grub config

3. removed the temporary hack to have one esp partition

# 🤔 Rationale
<!-- (can be skipped for bug fixes)
Why is this PR needed?
Reviewers should get a clear understanding of the goals of this PR.
-->

# 📝 Checks
<!-- Just a quick reminder :) -->

- [X] Format Check
- [X] Check rust
- [X] Updated API docs
- [X] Tests Pass
- [X] Functional Tests Pass
- [X] Local make targets work
  - [X] `make rpm`
  - [X] `make docker-build`

# 📌 Follow-ups
<!-- When applicable, add links to tasks or PR for work that will be addressed
later as a follow up of this PR. -->

- #0000

# 🗒️ Notes
<!-- Add any other relevant info -->

Related work items: #6584, #6601