Merged PR 17945: bug: Fix FT test_reinitialize_volume_fs()
<!-- Please, remember to prefix the title with one of:
 - `bug:`
 - `feature:`
 - `engineering:`
 - `docs:`.
For details, check [Pull Requests](/dev-docs/pull-requests.md) and
[this page](https://dev.azure.com/mariner-org/ECF/_git/ecf-docs?version=GBmain&path=/TeamDocs/dev-guide.md&_a=preview). -->

# 🔍 Description
<!--
What is this PR about?
Reviewers should get a clear understanding of what is being done/changed.
--> This PR fixes a bug in the functional test test_reinitialize_volume_fs(), by not relying on the existence of /dev/sdb and instead, setting up a new loop device.

# 🤔 Rationale
<!-- (can be skipped for bug fixes)
Why is this PR needed?
Reviewers should get a clear understanding of the goals of this PR.
-->

# 📝 Checks
<!-- Note any manual checks you've run -->

# 📌 Follow-ups
<!-- When applicable, add links to tasks or PR for work that will be addressed
later as a follow up of this PR. -->

- #0000

# 🗒️ Notes
<!-- Add any other relevant info -->

Fixed test_reinitialize_volume_fs()

Related work items: #6809