Merged PR 17959: bug: formatting of mounted unimaged encrypted volumes are skipped
Formatting of mounted unimaged encrypted volumes are skipped because the mount points in the host status are being initialized late.

The fix is to initialize the mount points in the host status before provisioning so that mount points can be detected by reading the host status. Also add a check to special-case swap mounts.

Related work items: #6831