Merged PR 17979: engineering: MIC API Alignment raw-zstd -> raw-zst
<!-- Please, remember to prefix the title with one of:
 - `bug:`
 - `feature:`
 - `engineering:`
 - `docs:`.
For details, check [Pull Requests](/dev-docs/pull-requests.md) and
[this page](https://dev.azure.com/mariner-org/ECF/_git/ecf-docs?version=GBmain&path=/TeamDocs/dev-guide.md&_a=preview). -->

# 🔍 Description
<!--
What is this PR about?
Reviewers should get a clear understanding of what is being done/changed.
-->

Change `raw-zstd` to `raw-zst`.

# 🤔 Rationale
<!-- (can be skipped for bug fixes)
Why is this PR needed?
Reviewers should get a clear understanding of the goals of this PR.
-->

Align with MIC to use the same name for image formats.

# 📝 Checks
<!-- Note any manual checks you've run -->

# 📌 Follow-ups
<!-- When applicable, add links to tasks or PR for work that will be addressed
later as a follow up of this PR. -->

# 🗒️ Notes
<!-- Add any other relevant info -->

Related work items: #6852