Merged PR 18009: engineering: Defining build purpose and testing environments for trident E2E pipeline
# 🔍 Description
Now we can select the different trident configurations to try according to the deployment environment for testing (Bare Metal or Virtual Machine) and the build purpose of the E2E pipeline (daily run to publish, changes in resources to main, triggered for pullrequest, or just for validations as manual run of the pipeline).

# 🤔 Rationale
We are starting to introduce trident configurations that may not be so stable in all deployment environments. Also, we are adding more and more configurations that we may not want to test all at once for each run of the pipeline.

# 📝 Checks
The following pipeline shows the changes are also working for BareMetal:
https://dev.azure.com/mariner-org/ECF/_build/results?buildId=502950&view=results

# 📌 Follow-ups
This PR will also serve to define the build purpose that will be used for this PR: https://dev.azure.com/mariner-org/ECF/_git/trident/pullrequest/17938

Related work items: #5966