Merged PR 18017: engineering: rename grubnoprefix efi to grub
<!-- Please, remember to prefix the title with one of:
 - `bug:`
 - `feature:`
 - `engineering:`
 - `docs:`.
For details, check [Pull Requests](/dev-docs/pull-requests.md) and
[this page](https://dev.azure.com/mariner-org/ECF/_git/ecf-docs?version=GBmain&path=/TeamDocs/dev-guide.md&_a=preview). -->

# 🔍 Description
<!--
What is this PR about?
Reviewers should get a clear understanding of what is being done/changed.
-->
rename the grub-noprefix efi to grub.efi
as M3 will switch over the main grub2 package to the no-prefix version by default and place a trampoline grub.cfg in the same folder as the EFI binary that can point to $prefix/grub.cfg.

# 🤔 Rationale
<!-- (can be skipped for bug fixes)
Why is this PR needed?
Reviewers should get a clear understanding of the goals of this PR.
-->

# 📝 Checks
<!-- Note any manual checks you've run -->

# 📌 Follow-ups
<!-- When applicable, add links to tasks or PR for work that will be addressed
later as a follow up of this PR. -->

- #0000

# 🗒️ Notes
<!-- Add any other relevant info -->

rename grubnpprefix to grub

Related work items: #6641