Merged PR 17938: engineering: E2E tests, select tests by trident configuration and purpose.
# 🔍 Description
With this PR we can now define which test we want to run in the pipeline according to the trident configuration and purpose.

The tests that will run are defined in a yaml file called `test-selection` that is located in the same directory than the respective trident configuration.  The following are the different categories for selecting tests and in general it is expected (but not required) that:
**compatible ⊆ weekly ⊆ daily ⊆ post_merge ⊆ pullrequest ⊆ validation**

    compatible: Tests that are compatible with a configuration
    weekly: Tests that run weekly in the pipeline
    daily: Tests that run nightly in the pipeline
    post_merge: Tests that run when the pipeline was triggered by a change in the resources
    pullrequest: Tests that run in a pipeline for a pull request
    validation: Tests that run when the pipeline was manually triggered.

Each group is defined by the previous one and we can select to add or remove tests or markers (groups of tests with the same label) in each level. This is an easy way to add and remove tests without having to check each of test implementation.

## Example of valid files to define the tests:
```
 compatible:
   - marker1
```
```
 compatible:
   - marker1
   - marker2
   - marker3
 daily:
   remove:
   - marker3
 pullrequest:
   remove:
   - file2.py::function_name1
validation:
   remove:
   - marker1
```

```
 compatible:
   - marker1
   - marker2
   - marker3
   - marker4
   - marker5
 weekly:
   remove:
   - file2.py::function_name1
 daily:
   remove:
   - marker5
 post_merge:
   remove:
   - marker4
   - marker2
   add:
   -  file2.py::function_name1
 pullrequest:
   remove:
   - marker3
   - file2.py::function_name1
 validation:
   remove:
   - marker1
   add:
   - file1.py::function_name1
   - file1.py::function_name2
```

# For review

The main changes are in the last two functions of `conftest.py` and `pytest.ini`. All the `test_selection.yaml` are the same as right now we just have a few basic tests. Everything inside of `helpers` is the same, the folder was just moved.

# 🤔 Rationale
Not all configurations will be compatible with all tests and some tests will take too much time to be run every time. Because of this, we need to be able to define which e2e tests will be run depending on the situation.

# 📌 Follow-ups
(https://dev.azure.com/mariner-org/ECF/_git/trident/pullrequest/18009)
(PR for documentation)

Related work items: #5966