Merged PR 18093: engineering: Use upstream mkfs directly
# 🔍 Description
We have been using systemd formatting logic, alas the error reporting is not great and it also does not fit well with containers and the rest of our workflow. Switching to using plan mkfs and mkswap to unblock. In the following changes, I will also replace growfs and the mounting logic.

Builds upon changes from Nadiia and Vince, but generalizes for all mounting needs.

# 🤔 Rationale
Simplify for more flexbility.

Related work items: #6650