Merged PR 18140: engineering: Moving the Trident BM tests from one BM Host (10.8.3.19) to another (10.8.3.18)
<!-- Please, remember to prefix the title with one of:
 - `bug:`
 - `feature:`
 - `engineering:`
 - `docs:`.
For details, check [Pull Requests](/dev-docs/pull-requests.md) and
[this page](https://dev.azure.com/mariner-org/ECF/_git/ecf-docs?version=GBmain&path=/TeamDocs/dev-guide.md&_a=preview). -->

# 🔍 Description
<!--
What is this PR about?
Reviewers should get a clear understanding of what is being done/changed.
-->

Moving `trident-Official` to run from one BM host to running on another BM host. Both hosts are Dell iDRAC and are identical in terms of specifications. Both are in AEP-02 lab.

# 🤔 Rationale
<!-- (can be skipped for bug fixes)
Why is this PR needed?
Reviewers should get a clear understanding of the goals of this PR.
-->

There was a lab rebuild on 16th Feb (long weekend of 19th) and the BM host ending with `.19` has not come back up successfully ever since. Successive rebuilds of the cluster on AEP02 has not fixed this. This is leading to random failures seen in the BM Pipeline runs.

Moving to `.18` is consistently giving successes and we should look to use this instead until we are able to understand further as to what broke on `.19`

# 📝 Checks
<!-- Note any manual checks you've run, base list of tests: dev-docs/manual-validation.md -->

Run: https://dev.azure.com/mariner-org/ECF/_build/results?buildId=511796&view=results

# 📌 Follow-ups
<!-- When applicable, add links to tasks or PR for work that will be addressed
later as a follow up of this PR. -->

# 🗒️ Notes
<!-- Add any other relevant info -->

Related work items: #6946