Merged PR 18144: Increase time for the job
# 🔍 Description
Increasing time for pipeline jobs that run the trident E2E tests from 15 to 20 minutes.

# 🤔 Rationale
Most of these jobs takes less than 15 minutes, but OneBranch tasks increase the time of the job. One of the jobs, is failing to finish by just a couple of minutes. The new time should solve this issue.

Related work items: #6951, #6952