Merged PR 18163: feature: Remove /host chroot from trident
# 🔍 Description
Eliminates the /host chroot from Trident when running as a container. This will allow for simpler logic in Trident.

# 🤔 Rationale
EMU and other dependencies of Trident had to be present in the host image, which caused additional problems with verity images and in general with RPM dependency tracking. Removing the need to chroot will simplify the flow.

Related work items: #6650, #6869