Merged PR 18218: bug: Removing the duplicate setting of default source as ado in trident e2e tests.
<!-- Please, remember to prefix the title with one of:
 - `bug:`
 - `feature:`
 - `engineering:`
 - `docs:`.
For details, check [Pull Requests](/dev-docs/pull-requests.md) and
[this page](https://dev.azure.com/mariner-org/ECF/_git/ecf-docs?version=GBmain&path=/TeamDocs/dev-guide.md&_a=preview). -->

#7011

# 🔍 Description
<!--
What is this PR about?
Reviewers should get a clear understanding of what is being done/changed.
-->

In order to address #6995 , there was a duplicate setting of `poetry source add` which was added which led to the error -
```
+ sudo env PATH=/home/cloudtest/.local/bin:/snap/bin:/home/cloudtest/.local/bin:/opt/pipx_bin:/home/cloudtest/.cargo/bin:/home/cloudtest/.config/composer/vendor/bin:/usr/local/.ghcup/bin:/home/cloudtest/.dotnet/tools:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/games:/usr/local/games:/snap/bin poetry source add --priority=default ado https://pkgs.dev.azure.com/mariner-org/ECF/_packaging/Mariner-HCI-Build-Dependencies-Feed/pypi/simple
Warning: Priority 'default' is deprecated. You can achieve the same effect by changing the priority to 'primary' and putting the source first.
Warning: Found deprecated priority 'default' for source 'ado' in pyproject.toml. You can achieve the same effect by changing the priority to 'primary' and putting the source first.
Source with name ado is already set to default. Only one default source can be configured at a time.
```

This is fixed by only setting the default source once and not calling to set it again later in the code.

# 🤔 Rationale
<!-- (can be skipped for bug fixes)
Why is this PR needed?
Reviewers should get a clear understanding of the goals of this PR.
-->

# 📝 Checks
<!-- Note any manual checks you've run, base list of tests: dev-docs/manual-validation.md -->

Pipeline Run on `trident`: https://dev.azure.com/mariner-org/ECF/_build/results?buildId=517987&view=results

# 📌 Follow-ups
<!-- When applicable, add links to tasks or PR for work that will be addressed
later as a follow up of this PR. -->

- Ideally we should be using a single template across `k8s-tests` and `trident` so as to apply all changes across repos. Need to clean this up to templatize the Test Setup code so that it can be reused across repositories `k8s-tests` and `trident`. This is of lower priority and shall be taken up in the future.

# 🗒️ Notes
<!-- Add any other relevant info -->

Related work items: #7011